Reorganized pipeline code
