+++ src/main.rs
@@ -1,14 +1,10 @@
 // Dependencies
 use std::{
     time::Duration,
-    sync::{Arc, Mutex},
-    collections::HashMap
+    sync::{Arc, Mutex}
 };
 use anyhow::Error;
 use serde_json::Value;
-use futures::future::join_all;
-use tokio::select;
-use tokio::sync::mpsc::unbounded_channel;
 
 // Project Modules
 mod constants;
@@ -17,6 +13,7 @@ mod error;
 mod logger;
 mod model;
 mod system_cmd;
+mod pipeline_manager;
 mod procedure_manager;
 mod util;
 
@@ -29,9 +26,9 @@ use model::{
     channel::PipelineConnection,
     project::pipeline::Condition
 };
-use system_cmd::{get_remote_git_repository_commits, setup_git_repository};
-use procedure_manager::run_procedure;
+use system_cmd::get_remote_git_repository_commits;
 use logger::LOGGER;
+use pipeline_manager::run_pipeline;
 use util::filesystem::read_configuration;
 
 #[cfg(feature = "http-api")]
@@ -115,7 +112,7 @@ async fn setup_updater(configuration: Arc<Mutex<Configuration>>, procedure_threa
                                             }
                 
                                             if !pipeline.stages.is_empty() {
-                                                let (pipeline_connection, mut receiver) = PipelineConnection::new(project.url.clone(), branch.name.clone(), pipeline.name.clone());
+                                                let (pipeline_connection, receiver) = PipelineConnection::new(project.url.clone(), branch.name.clone(), pipeline.name.clone());
                                                 pipeline_connections.push(pipeline_connection);
                                                 let default_deploy_path = configuration.default_deploy_path.clone();
                                                 let project_url = project.url.clone();
@@ -123,60 +120,7 @@ async fn setup_updater(configuration: Arc<Mutex<Configuration>>, procedure_threa
                                                 let short_hash = Arc::clone(&short_hash);
                                                 let default_log_path = configuration.default_log_path.clone(); // TODO: Revisit possible unnecessary clone (along with all its uses)
                 
-                                                tokio::task::spawn(async move {
-                                                    if let Ok((path, repository_name)) = setup_git_repository(&project_url, pipeline.deploy_path.as_ref().unwrap_or(&default_deploy_path), &pipeline.name, &pipeline.branches[branch_index]).await {
-                                                        let path = Arc::new(path);
-                                                        let default_log_path = Arc::new(format!("{}/{}", default_log_path, repository_name));
-                                                        for (stage_index, stage) in pipeline.stages.iter().enumerate() {
-                                                            if let Some(procedures) = Arc::clone(&pipeline).procedures.get(stage) {        
-                                                                let mut procedures_connection = HashMap::new();
-                                                                let mut procedures_handle = Vec::new();
-                                                                for procedure in procedures {
-                                                                    let (sender, receiver) = unbounded_channel();
-                                                                    let connection_id = match &procedure.name {
-                                                                        Some(procedure_name) => procedure_name.clone(),
-                                                                        None => pipeline.name.clone()
-                                                                    };
-
-                                                                    procedures_connection.insert(connection_id, sender);
-                                                                    let procedure_future = run_procedure(Arc::clone(&path), Arc::clone(&pipeline), stage_index, branch_index, Arc::clone(&short_hash), procedure.clone(), Arc::clone(&default_log_path), receiver);
-                                                                    procedures_handle.push(tokio::task::spawn(procedure_future));
-                                                                }
-
-                                                                select! {
-                                                                    procedure_results = join_all(procedures_handle) => {
-                                                                        for result in procedure_results {
-                                                                            if result.is_err() || result.unwrap().is_err() {
-                                                                                return;
-                                                                            }
-                                                                        }
-
-                                                                        info!(format!("[{}] [{}] Pipeline finished stage.", pipeline.name, stage));
-                                                                    },
-                                                                    Some(command) = receiver.recv() => {
-                                                                        match command {
-                                                                            Command::KillProcedure => {
-                                                                                debug!(format!("[{}] Pipeline kill command received. Dropping connections and ending task(s).", pipeline.name));
-                                                                                for (connection_id, sender) in &procedures_connection {
-                                                                                    if sender.send(Command::KillProcedure).is_err() {
-                                                                                        error!(format!("[{}] Pipeline failed to kill procedure with ID: {}. Continuing anyway.", pipeline.name, connection_id));
-                                                                                    }
-                                                                                }
-                                                                                break; // TODO: Re-evaluate sending command as dropping has same functionality
-                                                                            }
-                                                                        }
-                                                                    }
-                                                                }
-                                                            } else {
-                                                                error!(format!("[{}] Missing stage configuration! Stage: {}", pipeline.name, stage));
-                                                            }
-                                                        }
-
-                                                        info!(format!("[{}] Pipeline finished.", pipeline.name));
-                                                    } else {
-                                                        error!(format!("[{}] Failed to setup git repository. Skipping pipeline.", pipeline.name));
-                                                    }
-                                                });
+                                                tokio::task::spawn(run_pipeline(receiver, project_url, default_deploy_path, pipeline, short_hash, branch_index, default_log_path));
                                             }
                                         }
                                     } else {
